# openapi.yaml
openapi: 3.1.0
info:
  title: Dental Lab CRM API
  version: 0.2.0-mock
  summary: API-only mock contract for web and future mobile clients
  description: |
    Draft Swagger/OpenAPI specification for `dental-lab-crm`.
    The current runtime is API-only demo mode: auth is bypassed, Prisma and
    Supabase are disconnected, and mutations are in-memory for the active
    server process.

    Reference entities used by cases are archived instead of hard-deleted.
    Archived references are hidden from default management lists and creation
    option/dropdown endpoints, but historical case reads still include related
    archived records.
servers:
  - url: http://localhost:3000
    description: Local development
security: []
...
paths:
...
  /api/service-types:
    get:
      tags: [Registry]
      summary: List service types
      description: Returns active, unarchived service types for the logged-in user's dental lab, including each service type's workflowJson template.
      operationId: listServiceTypes
      responses:
        "200":
          description: Service types returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTypeListResponse"
    post:
      tags: [Registry]
      summary: Create service type
      description: Creates a service type for the logged-in user's dental lab. Optional `workflowJson.steps` rows must reference active processes in the same lab and have valid dependency ids.
      operationId: createServiceType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTypePayload"
      responses:
        "201":
          description: Service type created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTypeMutationResponse"
  /api/processes:
    get:
      tags: [Processes]
      summary: List processes
      description: Returns active, unarchived reusable lab process definitions for the logged-in user's dental lab.
      operationId: listProcesses
      responses:
        "200":
          description: Processes returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessListResponse"
    post:
      tags: [Processes]
      summary: Create process
      description: Creates a reusable lab process definition for the logged-in user's dental lab.
      operationId: createProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessPayload"
      responses:
        "201":
          description: Process created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessMutationResponse"
...
  /api/service-types/{id}:
    patch:
      tags: [Registry]
      summary: Update service type
      description: Updates a service type by id. When `workflowJson` is present, it replaces the service type's workflow template.
      operationId: updateServiceType
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTypePayload"
      responses:
        "200":
          description: Service type updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTypeMutationResponse"
    delete:
      tags: [Registry]
      summary: Archive service type
      description: Archives a service type by setting `isActive=false` and `deletedAt` to the current timestamp.
      operationId: archiveServiceType
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service type archived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceTypeMutationResponse"
  /api/processes/{id}:
    patch:
      tags: [Processes]
      summary: Update process
      description: Updates a process by id.
      operationId: updateProcess
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessPayload"
      responses:
        "200":
          description: Process updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessMutationResponse"
    delete:
      tags: [Processes]
      summary: Archive process
      description: Archives a process by setting `isActive=false` and `deletedAt` to the current timestamp. Existing case process snapshots keep their copied process names.
      operationId: archiveProcess
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process archived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessMutationResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServiceTypeSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          example: service_001
        name:
          type: string
          example: Crown
        workflowJson:
          $ref: "#/components/schemas/ServiceTypeWorkflow"
    Process:
      type: object
      required: [id, dentalLabId, name, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: process_001
        dentalLabId:
          type: string
          example: lab_001
        name:
          type: string
          example: Design
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
          example: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServiceTypeWorkflowStep:
      type: object
      required: [id, processId, dependsOn]
      properties:
        id:
          type: string
          example: design
        processId:
          type: string
          example: process_001
        dependsOn:
          type: array
          items:
            type: string
          example: []
    ServiceTypeWorkflow:
      type: object
      required: [steps]
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTypeWorkflowStep"
...
    ProcessPayload:
      type: object
      properties:
        name:
          type: string
          example: Design
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
    ServiceTypePayload:
      type: object
      properties:
        name:
          type: string
          example: Crown
        notes:
          type: string
          nullable: true
        isActive:
          type: boolean
        workflowJson:
          $ref: "#/components/schemas/ServiceTypeWorkflow"
...
    ProcessListResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Process"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    ProcessMutationResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/Process"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    ServiceTypeListResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTypeSummary"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    ServiceTypeMutationResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/ServiceTypeSummary"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
...